+++ cfgov/unprocessed/js/molecules/Expandable.js
@@ -1,9 +1,7 @@
 'use strict';
 
-// Required polyfills for <IE9.
-require( '../modules/polyfill/query-selector' );
-require( '../modules/polyfill/event-listener' );
-require( '../modules/polyfill/class-list' );
+// Required polyfills for IE9.
+if ( !Modernizr.classlist ) { require( '../modules/polyfill/class-list' ); } // eslint-disable-line no-undef, global-require, no-inline-comments, max-len
 
 // Required modules.
 var EventObserver = require( '../modules/util/EventObserver' );
@@ -75,7 +73,40 @@ function Expandable( element ) { // eslint-disable-line max-statements, inline-c
     if ( window.addEventListener ) {
       window.addEventListener( 'resize', _refreshHeight );
     }
+
+    _initObserver();
+
     return this;
+  }
+
+  /**
+   * Watch for the insertion/removal of DOM nodes.
+   * @returns {Object} The Expandable instance.
+   */
+  function _initObserver() {
+    var MutationObserver = window.MutationObserver ||
+                           window.WebKitMutationObserver ||
+                           window.MozMutationObserver;
+    var observeDOM;
+
+    if ( MutationObserver ) {
+      observeDOM = function() {
+        var observer = new MutationObserver( function( mutations ) {
+          mutations.forEach( _refreshHeight );
+        } );
+
+        observer.observe( _content, { childList: true, subtree: true } );
+      };
+    } else {
+      observeDOM = function() {
+        _content.addEventListener( 'DOMNodeInserted', _refreshHeight, false );
+        _content.addEventListener( 'DOMNodeRemoved', _refreshHeight, false );
+      };
+    }
+
+    window.setTimeout( observeDOM, 0 );
+
+    return _that;
   }
 
   /**