+++ cfgov/v1/forms.py
@@ -1,29 +1,31 @@
 import time
 from datetime import timedelta
+from util import ERROR_MESSAGES
 
 from django import forms
+from django.db.models import Q
 from django.forms.utils import ErrorList
-from django.forms.widgets import DateInput, SelectMultiple
 from django.contrib.auth import get_user_model
 from django.utils import timezone
-from django.db.models import Count
 from django.core.exceptions import ValidationError, ObjectDoesNotExist
-from django.contrib.auth import authenticate, get_user_model
+from django.contrib.auth import authenticate
 from django.conf import settings
+from django.forms import widgets
 
-from sheerlike.query import QueryFinder
 from sheerlike.templates import date_formatter
-from v1.models.events import EventPage
-from util import ERROR_MESSAGES
+from .models import ref
+from .models.learn_page import AbstractFilterPage
+from .util.util import most_common
+
 
 # importing wagtail.wagtailadmin.forms at module load time seems to have
 # caused some trouble with the translation subsystem being invoked before
 # it's ready-- so I've wrapped it in a function
 def login_form():
     from wagtail.wagtailadmin import forms as wagtail_adminforms
-    from v1.util import password_policy
 
     from .models import base
+
     class LoginForm(wagtail_adminforms.LoginForm):
 
         def clean(self):
@@ -32,24 +34,29 @@ def login_form():
 
             if username and password:
                 self.user_cache = authenticate(username=username,
-                  password=password)
+                                               password=password)
 
                 if (self.user_cache is None and username is not None):
                     UserModel = get_user_model()
 
-                    try: 
-                        user = UserModel._default_manager.get(username=username)
+                    try:
+                        user = UserModel._default_manager.get(
+                                username=username)
                     except ObjectDoesNotExist:
                         raise forms.ValidationError(
                             self.error_messages['invalid_login'],
                             code='invalid_login',
-                            params={'username': self.username_field.verbose_name},
+                            params={'username':
+                                    self.username_field.verbose_name
+                                    },
                         )
 
-                    # fail fast if user is already blocked for some other reason
-                    self.confirm_login_allowed(user) 
+                    # fail fast if user is already blocked for some other
+                    # reason
+                    self.confirm_login_allowed(user)
 
-                    fa, created = base.FailedLoginAttempt.objects.get_or_create(user=user)
+                    fa, created = base.FailedLoginAttempt.objects.\
+                        get_or_create(user=user)
                     now = time.time()
                     fa.failed(now)
                     # Defaults to a 2 hour lockout for a user
@@ -98,14 +105,16 @@ class FilterErrorList(ErrorList):
     def __str__(self):
         return '\n'.join(str(e) for e in self)
 
+
 class FilterDateField(forms.DateField):
-    def to_python(self, value):
+    def clean(self, value):
         if value:
             try:
                 value = date_formatter(value)
             except Exception as e:
                 pass
-        return super(FilterDateField, self).to_python(value)
+        return value
+
 
 class FilterCheckboxList(forms.CharField):
     def validate(self, value):
@@ -114,6 +123,7 @@ class FilterCheckboxList(forms.CharField):
             if self.label and '%s' in msg:
                 msg = msg % self.label
             raise forms.ValidationError(msg, code='required')
+
 
 class CalenderPDFFilterForm(forms.Form):
     filter_calendar = FilterCheckboxList(label='Calendar',
@@ -141,42 +151,154 @@ class CalenderPDFFilterForm(forms.Form):
             raise forms.ValidationError(ERROR_MESSAGES['DATE_ERRORS']['one_required'])
         return cleaned_data
 
-class EventsFilterForm(forms.Form):
-    tags_select_attrs = {
+
+class FilterableListForm(forms.Form):
+    topics_select_attrs = {
         'class': 'chosen-select',
         'multiple': 'multiple',
         'data-placeholder': 'Search for topics',
     }
+    authors_select_attrs = {
+        'class': 'chosen-select',
+        'multiple': 'multiple',
+        'data-placeholder': 'Search for authors',
+    }
     from_select_attrs = {
         'class': 'js-filter_range-date js-filter_range-date__gte',
         'type': 'text',
-        'placeholder': 'YYYY-MM',
+        'placeholder': 'dd/mm/yyyy',
     }
     to_select_attrs = from_select_attrs.copy()
     to_select_attrs.update({
         'class': 'js-filter_range-date js-filter_range-date__lte',
     })
 
-    from_date = forms.DateField(required=False, input_formats=['%Y-%m-%d'],
-                                widget=DateInput(attrs=from_select_attrs))
-    to_date = forms.DateField(required=False, input_formats=['%Y-%m-%d'],
-                              widget=DateInput(attrs=to_select_attrs))
-    topics = forms.MultipleChoiceField(required=False, choices=[],
-                                       widget=SelectMultiple(attrs=tags_select_attrs))
-    @property
-    def field_queries(self):
-        return zip(['start_dt__gte', 'end_dt__lte', 'tags__name__in'],
-                   ['from_date', 'to_date', 'topics'])
+    title = forms.CharField(max_length=250, required=False)
+    from_date = forms.DateField(
+        required=False,
+        input_formats=['%d/%m/%Y'],
+        widget=widgets.DateInput(attrs=from_select_attrs))
+    to_date = forms.DateField(
+        required=False,
+        input_formats=['%d/%m/%Y'],
+        widget=widgets.DateInput(attrs=to_select_attrs))
+    categories = forms.MultipleChoiceField(
+        required=False,
+        choices=ref.page_type_choices,
+        widget=widgets.CheckboxSelectMultiple())
+    topics = forms.MultipleChoiceField(
+        required=False,
+        choices=[],
+        widget=widgets.SelectMultiple(attrs=topics_select_attrs))
+    authors = forms.MultipleChoiceField(
+        required=False,
+        choices=[],
+        widget=widgets.SelectMultiple(attrs=authors_select_attrs))
 
     def __init__(self, *args, **kwargs):
-        super(EventsFilterForm, self).__init__(*args, **kwargs)
+        parent = kwargs.pop('parent')
+        super(FilterableListForm, self).__init__(*args, **kwargs)
+        self.set_topics(parent)
+        self.set_authors(parent)
 
-        if 'topics' in self.fields:
-            options = list(set([tag for tags in [event.tags.names() for event
-                                                 in EventPage.objects.live()]
-                                for tag in tags]))
-            most = [(option, option) for option in options[:3]]
-            other = [(option, option) for option in options[3:]]
+    # Populate Topics' choices
+    def set_topics(self, parent):
+        all_tags = [tag for tags in [page.tags.names() for page in
+                    AbstractFilterPage.objects.live().descendant_of(
+                    parent).live()] for tag in tags]
+        # Orders by most to least common tags
+        options = most_common(all_tags)
+        most = [(option, option) for option in options[:3]]
+        other = [(option, option) for option in options[3:]]
+        self.fields['topics'].choices = \
+            (('Most frequent', tuple(most)),
+             ('All other topics', tuple(other)))
 
-            self.fields['topics'].choices = (('Most frequent', tuple(most)),
-                                             ('All other topics', tuple(other)))
+    # Populate Authors' choices
+    def set_authors(self, parent):
+        all_authors = [author for authors in [page.authors.names() for page in
+                       AbstractFilterPage.objects.live().descendant_of(
+                       parent).live()] for author in authors]
+        # Orders by most to least common authors
+        self.fields['authors'].choices = [(author, author) for author in
+                                          most_common(all_authors)]
+
+    def clean(self):
+        cleaned_data = super(FilterableListForm, self).clean()
+        from_date = cleaned_data.get('from_date')
+        to_date = cleaned_data.get('to_date')
+        # Check if both date_lte and date_gte are present
+        # If the 'start' date is after the 'end' date, swap them
+        if (from_date and to_date) and to_date < from_date:
+            data = dict(self.data)
+            data_to_date = data['to_date']
+            self.cleaned_data['to_date'], data['to_date'] = \
+                from_date, data['from_date']
+            self.cleaned_data['from_date'], data['from_date'] = \
+                to_date, data_to_date
+            self.data = data
+        return self.cleaned_data
+
+    # Does the job of {{ field }}
+    # In the template, you pass the field and the id and name you'd like to
+    # render the field with.
+    def render_with_id(self, field, attr_id):
+        for f in self.fields:
+            if field.html_name in f:
+                self.fields[f].widget.attrs.update({'id': attr_id})
+                self.set_field_html_name(self.fields[f], attr_id)
+                return self[f]
+
+    # Sets the html name by replacing the render method to use the given name.
+    def set_field_html_name(self, field, new_name):
+        """
+        This creates wrapper around the normal widget rendering,
+        allowing for a custom field name (new_name).
+        """
+        old_render = field.widget.render
+        if isinstance(field.widget, widgets.SelectMultiple):
+            field.widget.render = lambda name, value, attrs=None, choices=(): \
+                old_render(new_name, value, attrs, choices)
+        else:
+            field.widget.render = lambda name, value, attrs=None: \
+                old_render(new_name, value, attrs)
+
+    # Generates a query by iterating over the zipped collection of
+    # tuples.
+    def generate_query(self):
+        final_query = Q()
+        if self.is_bound:
+            for query, field_name in zip(self.get_query_strings(), self.declared_fields):
+                if self.cleaned_data.get(field_name):
+                    final_query &= \
+                        Q((query, self.cleaned_data.get(field_name)))
+        return final_query
+
+    # Returns the field to order the list by
+    def get_order_attr(self):
+        return 'date_published'
+
+    # Returns a list of query strings to associate for each field, ordered by
+    # the field declaration for the form. Note: THEY MUST BE ORDERED IN THE
+    # SAME WAY AS THEY ARE DECLARED.
+    def get_query_strings(self):
+        return [
+            'title__icontains',      # title
+            'date_published__gte',   # from_date
+            'date_published__lte',   # to_date
+            'categories__name__in',  # categories
+            'tags__name__in',        # topics
+            'authors__name__in',     # authors
+        ]
+
+
+class EventArchiveFilterForm(FilterableListForm):
+    def get_query_strings(self):
+        return [
+            'title__icontains',      # title
+            'start_dt__gte',         # from_date
+            'end_dt__lte',           # to_date
+            'categories__name__in',  # categories
+            'tags__name__in',        # topics
+            'authors__name__in',     # authors
+        ]