+++ cfgov/v1/models/demo.py
@@ -26,7 +26,6 @@ class DemoPage(CFGOVPage):
         ('call_to_action', molecules.CallToAction()),
         ('expandable', molecules.Expandable()),
         ('featured_content', molecules.FeaturedContent()),
-        ('notification', molecules.Notification()),
 
     ], blank=True)